\name{catlearn-package}
\alias{catlearn}
\docType{package}
\title{
  
Formal Modeling for Psychology.

}
\description{

  Formal psychological models, independently-replicated data sets
  against which to test them, and simulation archives.

}

\details{

  For a complete list of functions, use \code{library(help = "catlearn")}.
  
  For a complete table of simulations, use \code{data(thegrid)}.
  
}


\author{
  
Andy Wills

Maintainer: Andy Wills \email{andy@willslab.co.uk}

}

\keyword{ package }
