\name{shj61}
\docType{data}
\alias{shj61}
\title{Nosofsky et al. (1994) results}
\description{
  The \code{shj61} data frame has 96 rows and 3 columns of data from an
  experiment on the effect of category structure on response accuracy
  in a category learning task. 
}
\usage{data(shj61)}
\format{
  A data frame with the following columns:
  \describe{
    \item{type}{Type of category structure, as defined by Shepard et al. (1961). 
    Takes values : 1-6}
    \item{block}{Training block. Takes values: 1-16}
    \item{error}{Mean error probability}
    }
}
\details{
  Mean error probability by training block for each of Shepard et al.'s (1961) six
  category structure types; from a replication by Nosofsky et al. (1994). Only the
  first 16 blocks are reported, for comparability with the model fitting reported
  in that paper.
}
\source{
  Nosofsky, R.M., Gluck, M.A., Plameri, T.J., McKinley, S.C. and Glauthier, P. 
  (1994). Comparing models of rule-based classification learning: A replication
  and extension of Shepaard, Hovland, and Jenkins (1961). \emph{Memory and Cognition, 
  22}, 352--369
  }

\examples{
}
\keyword{datasets}
