\name{ssecl}
\alias{ssecl}
\title{Sum of squared errors}
\description{
Calculate sum of squared errors
}
\usage{ssecl(obs,exp)}
\arguments{
  \item{obs}{Vector of observed values}
  \item{exp}{Vector of expected values}
}   
\value{
Returns sum of the squared differences.
}
\details{
}
\author{
Andy Wills
}
\references{
}
\seealso{
}
\examples{
}
