\name{catlearn-package}
\alias{catlearn}
\docType{package}
\title{
  
Formal Modeling for Psychology.

}
\description{

  Formal psychological models, independently-replicated data sets
  against which to test them, and simulation archives.

}

\details{

  For a complete list of functions, use \code{library(help =
    "catlearn")}.

  For a complete table of simulations, use \code{data(thegrid)}.

  All functions are concisely documented, use the help function e.g
  \code{?shin92}.
  
  For more detailed documentation, see the references listed by the help
  documentation.

  For a tutorial introduction, see Wills et al. (2016a).

  For a guide to contributing to this package, Catlearn Research Group
  (2016).
  
}


\author{
  
Andy Wills

Maintainer: Andy Wills \email{andy@willslab.co.uk}

}

\references{

  Catlearn Research Group (2016). Contributing to catlearn.
  \url{http://catlearn.r-forge.r-project.org/intro-catlearn.pdf}
  
  Wills, A.J., O'Connell, G., Edmunds, C.E.R. & Inkster, A.B. (2016).
  Progress in modeling through distributed collaboration: Concepts,
  tools, and category-learning examples. \emph{The Psychology of
  Learning and Motivation}.
  
}

\keyword{ package }
