\name{catlearn-package}
\alias{catlearn}
\docType{package}
\title{
  
CATalog of LEARNing models.

}
\description{
  
  Learning models, plus archives of classic phenomena (CIRP), input
  representations, and simulations.

}

\details{

  For a complete list of functions, use \code{library(help = "catlearn")}.
  
  For a complete table of simulations, use \code{data(thegrid)}.
  
}


\author{
  
Andy Wills

Maintainer: Andy Wills \email{andy@willslab.co.uk}

}

\keyword{ package }
