\name{catlearn-package}
\alias{catlearn}
\docType{package}
\title{
Computational models of category learning
}
\description{
Category learning models, implemented as list processors.
}
\details{

}
\author{
Andy Wills

Maintainer: Andy Wills \email{andy@wilslab.co.uk}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
