\name{catlearn-package}
\alias{catlearn}
\docType{package}
\title{
  
Computational Models of Category Learning

}
\description{
  
  Category learning models, implemented as list processors.
  
}
\details{

}
\author{
  
Andy Wills

Maintainer: Andy Wills \email{andy@willslab.co.uk}

}

\references{
  
}
\keyword{ package }
\seealso{
}
\examples{
}
