\name{stsimGCM}
\alias{stsimGCM}
\title{
  Similarity-Dissimilarity Generalized Context Model
}

\description{
  An extension to Nosofsky's (1984, 2011) Generalized Context Model ; an
  exemplar-based model of categorization with differentially weighed similarity
  and dissimilarity.
}

\usage{
  stsimGCM(st, test)
}
\arguments{
  \item{st}{List of model parameters}
}

\value{
  Returns a list of the following matrices:
  \code{Evidence} The evidence for each category
}

\details{

}

\author{ Lenard Dome, Andy Wills }

\references{


}

\examples{

}

\note{

  1.

  2.

  }
